airavata-mft-portal/airavata_mft_portal/requirements.txt
Django==3.0.4
djangorestframework==3.11.0
grpcio==1.27.2
googleapis-common-protos==1.51.0
